+++ src/pluginsettings.hpp
@@ -14,6 +14,7 @@ struct Settings {
 	bool mbFavoriteHighlight = true;
 	float mbSearchThreshold = 0.5f;
 	bool mbMagnifierEnabled = false;
+	bool mbApplyLibraryWhitelist = false;
 
 	NVGcolor overlayTextColor = bndGetTheme()->menuTheme.textColor;
 	int overlayHpos = 0;